Adjust Python sink to current Babeltrace API
Adjust the code to the actual API of Babeltrace 2.0.  I was able to run
the scenario 1 (the one that uses Python logging) successfully.  I'm
having trouble running the scenario 2, so I'll stop here for now.  This
means there are paths in the component code that are not converted yet,
and probably don't work.